+++ lib/script/Motion.java
@@ -12,7 +12,7 @@ public class Motion {
 	public Color colorto = Color.BLUE;	// change obj's color
 	public Point start = null;		// obj shape start: such as arrow's bottom
 	public Point end = null;		// obj shape end: such as arrow's head
-	public int angle = 0;			// obj shape starting angle
+	public double angle = 0;			// obj shape starting angle
 									// 0 degree: <--
 									// 90 degree: pointing down
 									// 180 degree: -->
